src/restaurant-info-server/requirements.txt
grpcio     
grpcio-tools
grpcio-health-checking
LatLon23